terraform: destroy module nodes show up in plan destroy
