+++ terraform/graph_builder_destroy_plan.go
@@ -44,6 +44,9 @@ func (b *DestroyPlanGraphBuilder) Steps() []GraphTransformer {
 		// Attach the configuration to any resources
 		&AttachResourceConfigTransformer{Module: b.Module},
 
+		// Module destroy nodes
+		&ModuleDestroyTransformer{State: b.State},
+
 		// Single root
 		&RootTransformer{},
 	}
